back/requirements.txt
psycopg2-binary==2.9.5
python-dotenv==0.21.1
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.14.0
mongoengine==0.26.0
pymongo==4.3.3
django-cors-headers==3.13.0
certifi==2020.12.5
chardet==4.0.0
oauthlib==3.1.0
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.4
#idna==3.1
